Add simple model
Model follows Pytorch "Text classification with the torchtext library" tutorial, which uses an
embedding bag layer and a linear layer.
The dataset have not been split into training and test data, currently training and testing on
the same dataset. The features used from the dataset to trained only uses the concatenation
of the featrues: company_profile, description and requirements, as a single feature to train
on.